pyproject.toml
[project]
name = "syntek-modules"
version = "0.1.0"
description = "Modular Django, React, React Native, and Rust modules for Syntek projects"
readme = "README.md"
requires-python = ">=3.14"
license = {text = "MIT"}
authors = [
    {name = "Syntek Dev", email = "dev@syntek.com"}
]
keywords = ["django", "react", "react-native", "rust", "modules"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.14",
    "Framework :: Django :: 6.0",
]

# Root project dependencies (for development/testing)
dependencies = [
    "django>=6.0.2",
    "strawberry-graphql>=0.291.3",
    "psycopg[binary]>=3.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.15.0",
    "pyright>=1.1.0",
    "django-stubs>=5.1.0",
    "ipython>=8.20.0",
    "pre-commit>=4.0.0",
    "detect-secrets>=1.5.0",
]
test = [
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "factory-boy>=3.3.0",
    "faker>=25.0.0",
]
docs = [
    "sphinx>=7.3.0",
    "sphinx-rtd-theme>=2.0.0",
    "sphinx-autodoc-typehints>=2.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend", "rust"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = [
    "tests",
    "backend/security-auth/authentication/tests",
]
addopts = "-v --tb=short --cov --cov-report=html --cov-report=term"

[tool.coverage.run]
source = ["backend"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "S",   # flake8-bandit (security)
    "A",   # flake8-builtins
    "DTZ", # flake8-datetimez
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "S101",  # use of assert (OK in tests)
    "S105",  # possible hardcoded password (often false positives)
    "S106",  # possible hardcoded password (often false positives)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports
"*/migrations/*.py" = ["E501"]  # long lines in migrations
".claude/plugins/*.py" = ["S", "S110", "S112", "S324", "S603", "S607", "S311"]  # exclude security checks for plugins

[tool.pyright]
pythonVersion = "3.14"
typeCheckingMode = "basic"
reportMissingImports = true
reportMissingTypeStubs = false
stubPath = ""
exclude = [
    "**/migrations/**",
    "**/tests/**",
    "graphql/examples/**",
    "examples/**",
    ".venv/**",
    "build/**",
    "dist/**"
]

[tool.django-stubs]
django_settings_module = "tests.settings"
strict_settings = false

[dependency-groups]
dev = [
    "maturin>=1.11.5",
]
